requirements.txt
flask

flask_material

flask_cors

pymongo

datetime

requests

mysql-connector-python==8.0.17
